Remove references to normal3, this is unsupported
There is a typo on http://www.pbrt.org/fileformat-v3.html#parameter-lists
which lists normal3 as a valid type which it is not.

Remove references to this as well as rely on the VOP nodes to have the
correct 'pbrt.type' name instead of auto-suffixing it with the size. This
requires the node creators to be little more careful but avoids behind the
scenes magic which isn't always obvious.

Also renamed pbrt_type to pbrt.type which is what it should be.