Made naming of instancing functions more clear
Originially the list_instances() function was supposed to be much
more emcompassing but after learning that the only cases where it
would be useful is Fast Instancing the naming and comments were a
bit misleading. The workflow is to look at a Houdini instance node
and report any references that node/geo makes to other geometries.