Change spectrum parm type separator to : from .
Previous we could specify the spectrum type in an override with .
At some point in the future when specifying overrides with node names
this will no longer work as . can be part of a node name. This is switched
to a : which is not allowed as part of a node name.