Allow for str spectrum values
The spectrum types support either a array of numerics or a file path
the change tests to see if it a string representation of numerics was
passed and then converted. This is to handle cases with the material
overrides where the existing nodes don't allow for passing advanced
parm types like in the Spectrum node.