Support for Full Point Instancing of Mediums
Primary change is the ability to now attach PBRT Mediums to points
along with point attributes that map to their parameters. To support
both interior and exterior attributes the various parameters are
prefixed with either interior or exterior. For example -
interior_sigma_a on a point will override whatever the value was fetched
from the pbrt_interior material.

As side benefit this means that volume primitives can now be varied per
instance as well. (This is useful since volumes can't have standard
material overrides.) To support this meta properties (starting with a .)
are now passed down the various wranglers. While material_overrides are a
string, the interior_override is actually a ParamSet.

As a note tho, instancing of volumes in PBRT isn't that useful in practice
due to lack of support of overlapping volumes, at least in pbrt v3.