Added a caching mechansim for Node overrides
The caching mechanism is to target cases where the same Material
Node is being reused many times to apply overrides. Instead of
reinterpreting the parameters for every override across many prims
cache those results and then apply just the values.

For a scene with 100,000 curves, the caching reduced the time of
PBRTnodes.override_paramset from 101sec to 14sec, with the majority
of the remaining time being evals().